requirements.txt
flask
pandas
jsonify
openai
python-dotenv